# Add `invokelatest` barrier to `string(...)` in `@assert`