@IanButterworth 2024-09-11
Great! This fixes the `Artifacts._artifact_str` invalidation seen here https://github.com/JuliaPackaging/LazyArtifacts.jl/pull/1#issuecomment-2341588541

### Master with https://github.com/JuliaPackaging/LazyArtifacts.jl/pull/1
```
% ./julia --start=no --trace-compile=stderr --trace-compile-timing -e "using InteractiveUtils; @time @time_imports using MicroMamba"
      0.5 ms  Printf
     15.0 ms  Dates
      0.3 ms  Scratch
     10.0 ms  LazyArtifacts
      1.0 ms  TOML
      7.2 ms  Preferences
      0.5 ms  JLLWrappers
               ┌ 545.2 ms micromamba_jll.__init__() 99.93% compilation time (88% recompilation)
    545.7 ms  micromamba_jll 99.84% compilation time (88% recompilation)
      0.4 ms  MicroMamba
  0.586889 seconds (763.19 k allocations: 40.931 MiB, 5.56% gc time, 92.84% compilation time: 88% of which was recompilation)
```
Full detail:
```
% ./julia --start=no --trace-compile=stderr --trace-compile-timing -e "using InteractiveUtils; @time @time_imports using MicroMamba"
      0.5 ms  Printf
     15.0 ms  Dates
      0.3 ms  Scratch
     10.0 ms  LazyArtifacts
      1.0 ms  TOML
      7.2 ms  Preferences
      0.5 ms  JLLWrappers
#=    0.0 ms =# precompile(Tuple{typeof(micromamba_jll.__init__)})
#=    0.0 ms =# precompile(Tuple{typeof(micromamba_jll.find_artifact_dir)})
#=    2.4 ms =# precompile(Tuple{Type{Base.Val{x} where x}, Module})
#=    2.0 ms =# precompile(Tuple{Type{NamedTuple{(:honor_overrides,), T} where T<:Tuple}, Tuple{Bool}})
#=  439.9 ms =# precompile(Tuple{typeof(Artifacts._artifact_str), Module, String, Base.SubString{String}, String, Base.Dict{String, Any}, Base.SHA1, Base.BinaryPlatforms.Platform, Base.Val{LazyArtifacts}})
#=   24.4 ms =# precompile(Tuple{typeof(Base.unique!), Array{String, 1}})
#=    1.3 ms =# precompile(Tuple{typeof(Base.invokelatest), Any})
#=    0.0 ms =# precompile(Tuple{typeof(JLLWrappers.get_julia_libpaths)})
#=    7.8 ms =# precompile(Tuple{typeof(Base.vcat), Array{String, 1}, Array{String, 1}})
               ┌ 545.2 ms micromamba_jll.__init__() 99.93% compilation time (88% recompilation)
    545.7 ms  micromamba_jll 99.84% compilation time (88% recompilation)
      0.4 ms  MicroMamba
#=    2.0 ms =# precompile(Tuple{Type{NamedTuple{(:value, :time, :bytes, :gctime, :gcstats, :lock_conflicts, :compile_time, :recompile_time), T} where T<:Tuple}, Tuple{Nothing, Float64, Int64, Float64, Base.GC_Diff, Int64, Float64, Float64}})
#=    2.9 ms =# precompile(Tuple{typeof(Base.getproperty), NamedTuple{(:value, :time, :bytes, :gctime, :gcstats, :lock_conflicts, :compile_time, :recompile_time), Tuple{Nothing, Float64, Int64, Float64, Base.GC_Diff, Int64, Float64, Float64}}, Symbol})
#=   26.4 ms =# precompile(Tuple{typeof(Core.kwcall), NamedTuple{(:msg,), Tuple{Nothing}}, typeof(Base.time_print), Base.TTY, Float64, Int64, Int64, Int64, Int64, Float64, Float64, Bool})
  0.586889 seconds (763.19 k allocations: 40.931 MiB, 5.56% gc time, 92.84% compilation time: 88% of which was recompilation)
  ```

### This PR with https://github.com/JuliaPackaging/LazyArtifacts.jl/pull/1
```
julia> @time @time_imports using MicroMamba
      0.8 ms  Printf
     18.3 ms  Dates
      0.4 ms  Scratch
     10.3 ms  LazyArtifacts
      0.6 ms  TOML
      6.8 ms  Preferences
      0.5 ms  JLLWrappers
               ┌ 3.9 ms micromamba_jll.__init__() 95.66% compilation time (37% recompilation)
      4.5 ms  micromamba_jll 82.93% compilation time (37% recompilation)
      0.4 ms  MicroMamba
  0.068322 seconds (68.55 k allocations: 3.725 MiB, 5.49% compilation time: 37% of which was recompilation)
```