@aviatesk 2024-09-12
I think this is a great idea. The 1-arg `@assert` often leads to poor type inference due to recursive extended inference (xref: JuliaLang/julia#55625). Using `invokelatest` would eliminate that concern too.