@vtjnash 2024-09-14
We could also use invokeinworld with tls_world_age, or an inferencebarrier to achieve similar effects?